+++ CHANGELOG.md
@@ -3,6 +3,12 @@ xcode CHANGELOG
 
 This file is used to list changes made in each version of the xcode cookbook.
 
+2.1.0
+-----
+
+- Modify app install for license and multi install (replaced with suffix)
+- Add ability to install simulator
+
 2.0.0
 -----